+++ src/clj/skylobby/fx/battle.clj
@@ -544,25 +544,22 @@
        (when (not singleplayer)
          [(let [am-away (:away my-client-status)]
             (merge
-              {:fx/type :button
-               :text (if am-away "Away" "Here")
-               :on-action {:event/type :spring-lobby/update-client-status
-                           :client-data (when-not singleplayer client-data)
-                           :client-status (assoc my-client-status :away (not am-away))}}
-              (when am-away
-                {:graphic
-                 {:fx/type font-icon/lifecycle
-                  :icon-literal "mdi-sleep:16:grey"}})))])
-       [{:fx/type :button
-         :text (if am-spec
-                 "Spectating"
-                 "Playing")
-         :on-action {:event/type :spring-lobby/battle-spectate-change
-                     :client-data (when-not singleplayer client-data)
-                     :is-me true
-                     :is-bot false
-                     :id my-player
-                     :value am-spec}}])}
+              {:fx/type :combo-box
+               :value (if am-away "Away" "Here")
+               :items ["Away" "Here"]
+               :on-value-changed {:event/type :spring-lobby/on-change-away
+                                  :client-data (when-not singleplayer client-data)
+                                  :client-status (assoc my-client-status :away (not am-away))}}))])
+       [{:fx/type :combo-box
+         :value (if am-spec
+                  "Spectating"
+                  "Playing")
+         :items ["Spectating" "Playing"]
+         :on-value-changed {:event/type :spring-lobby/on-change-spectate
+                            :client-data (when-not singleplayer client-data)
+                            :is-me true
+                            :is-bot false
+                            :id my-player}}])}
     {:fx/type :h-box
      :alignment :center-left
      :style {:-fx-font-size 24}
@@ -1063,7 +1060,8 @@
         engine-details (get engines-by-version engine-version)
         indexed-map (get maps-by-name battle-map)
         battle-map-details (resource/cached-details map-details indexed-map)
-        indexed-mod (get mods-by-name battle-modname)
+        indexed-mod (or (get mods-by-name battle-modname)
+                        (get mods-by-name (u/mod-name-git-no-ref battle-modname)))
         battle-mod-details (resource/cached-details mod-details indexed-mod)
         mod-dependencies (->> battle-modname
                               resource/mod-dependencies